requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
groq==0.9.0
requests==2.28.2
pydantic==1.10.7